Move type definitions into submodule; add basic support for gists
