+++ servant-github.cabal
@@ -37,6 +37,8 @@ library
                      , Network.GitHub.API
                      , Network.GitHub.Client
                      , Network.GitHub.Types
+                     , Network.GitHub.Types.Core
+                     , Network.GitHub.Types.Gist
   build-depends:       base >= 4.7 && < 5
                      , aeson
                      , bytestring
@@ -54,6 +56,7 @@ library
                      , transformers
                      , lens
                      , mtl
+                     , unordered-containers
   default-language:    Haskell2010
 
 executable test
